+++ AAA_API/classes/initialise.php
@@ -47,6 +47,7 @@ class Initialise extends Database {
 			Password		VARCHAR(256)	NOT NULL,
 			AccountStatus	INT(1)			NOT NULL,
 			AccessToken		VARCHAR(256),
+			SpotifyEmail	VARCAHR(250),
 			
 			PRIMARY KEY (ID),
 			UNIQUE(PublicID)
@@ -353,6 +354,55 @@ class Initialise extends Database {
 
 
 
+	/**
+	 * Creates PLAYLISTS table
+	 * 
+	 * @param		mysqli		The database to create the table in
+	 */
+	private static function createPlaylists(mysqli $conn) {
+
+		$tableName = "PLAYLISTS";
+		$SQL = "CREATE TABLE $tableName (
+			ID				INT(11)			NOT NULL	AUTO_INCREMENT,
+			PublicID		VARCHAR(32)		NOT NULL,
+			Name			VARCHAR(100)	NOT NULL,
+
+			PRIMARY KEY (ID),
+			UNIQUE(PublicID)
+		);";
+		$res = self::createTable($conn, $SQL, $tableName);
+
+	}
+
+
+
+
+
+	/**
+	 * Creates PLAYLISTS_TO_USERS table
+	 * 
+	 * @param		mysqli		The database to create the table in
+	 */
+	private static function createPlaylistsToUsers(mysqli $conn) {
+
+		$tableName = "PLAYLISTS_TO_USERS";
+		$SQL = "CREATE TABLE $tableName (
+			ID				INT(11)			NOT NULL	AUTO_INCREMENT,
+			PlaylistID		VARCHAR(50)		NOT NULL,
+			UserID			VARCHAR(32)		NOT NULL,
+
+			PRIMARY KEY (ID),
+			FOREIGN KEY (UserID) REFERENCES USERS (PublicID) ON DELETE CASCADE,
+			FOREIGN KEY (PlaylistID) REFERENCES PLAYLISTS (PublicID) ON DELETE CASCADE
+		);";
+		$res = self::createTable($conn, $SQL, $tableName);
+
+	}
+
+
+
+
+
 
 
 
@@ -379,6 +429,8 @@ class Initialise extends Database {
 		self::createTracksToUsers($conn);
 		self::createTracksToLabels($conn);
 		self::createRequests($conn);
+		self::createPlaylists($conn);
+		self::createPlaylistsToUsers($conn);
 
 		// Insert special rights into table