+++ AAA_WEBAPP/assets/js/api.js
@@ -160,7 +160,7 @@ Api.get = {
 	tracks: async () => {
 		const res = await Api.sendRequest("api/v1/tracks/get/", "GET");
 
-		Collection.tracks = [];
+		Collection.reset();
 		for(const track of res.data) {
 			const t = new Track(track.id, track.name, track.artists, new Date(track.addedAt), new Date(track.releaseDate), track.labels);
 			Collection.add(t);
@@ -197,7 +197,19 @@ Api.get = {
 		import: async () => {
 			const res = await Api.sendRequest("api/v1/spotify/playlists", "GET");
 			return (res.data ? res.data : res);
-		}
+		},
+
+
+
+		/**
+		 * Gets the labels from the database
+		 * 
+		 * @returns {array} The labels
+		 */
+		mine: async () => {
+			const res = await Api.sendRequest("api/v1/playlists/", "GET");
+			return (res.data ? res.data : res);
+		},
 
 	}
 
@@ -273,7 +285,7 @@ Api.show = {
 			const textContainer = Api.createElement("div", { classList: "text" });
 
 			textContainer.appendChild(Api.createElement("p", { innerHTML: label.name }));
-			textContainer.appendChild(Api.createElement("p", { innerHTML: "xx songs", classList: "right" }));
+			textContainer.appendChild(Api.createElement("p", { innerHTML: label.numTracks + " song" + (label.numTracks !== 1 ? "s" : ""), classList: "right" }));
 
 			row.appendChild(textContainer);
 
@@ -345,8 +357,71 @@ Api.show = {
 
 			}	// for list of playlists
 
-		}	// Api.show.playlists.import
+		},	// Api.show.playlists.import
 
-	}	// Api.show.playlists
+
+
+		/**
+		 * Displays the playlists
+		 * 
+		 * @param {array} playlists The playlists array
+		 */
+		mine: (playlists) => {
+	
+			// Clears the output and gets the data
+			const output = document.getElementById("playlists");
+			output.innerHTML = "";
+	
+			// For every row
+			for(const playlist of playlists) {
+	
+				// Create the row and text container
+				const row = Api.createElement("div", { classList: "row" });
+				const textContainer = Api.createElement("div", { classList: "text" });
+				textContainer.appendChild(Api.createElement("p", { innerHTML: playlist.name }));
+	
+				row.appendChild(textContainer);
+	
+				// Create edit button
+				row.appendChild(Api.createIcon("edit", () => {
+	
+					const popup = new BigPopup("Edit Playlist", "api/v1/playlists/" + playlist.publicID + "/update", "POST", "edit-playlist-form");
+					popup.add("input", "Name", { value: playlist.name });
+					popup.show("EDIT");
+					HtmlJsForm.findById("edit-playlist-form").addCallback(async () => { Api.show.playlists.mine(await Api.get.playlists.mine()); });
+	
+				}));
+	
+				// Create remove button
+				row.appendChild(Api.createIcon("delete", () => {
+	
+					const popup = new BigPopup("Remove Playlist", "api/v1/playlists/" + playlist.publicID + "/delete", "DELETE", "remove-playlist-form");
+					popup.add("p", "text", { innerHTML: "Are you sure you want to remove \"" + playlist.name + "\"? All songs affiliated with this playlist will lose their association, and it cannot be undone." });
+					popup.show("REMOVE");
+					HtmlJsForm.findById("remove-playlist-form").addCallback(async () => { Api.show.playlists.mine(await Api.get.playlists.mine()); });
+	
+				}));
+	
+				// append row to HTML
+				output.appendChild(row);
+	
+			}
+
+		}	// Api.show.playlists.mine
+
+	},	// Api.show.playlists
+
+
+
+
+
+	/**
+	 * Shows the email address of the user
+	 * 
+	 * @param {string} email The email address
+	 */
+	spotifyEmail(email) {
+		document.getElementById("spotify-email").innerHTML = email;
+	}
 
 }	// Api.show