+++ AAA_WEBAPP/assets/js/option-popup.js
@@ -65,7 +65,7 @@ class OptionPopup {
 			const removeLabel = Api.createElement("div", { classList: "row" });
 			removeLabel.addEventListener("click", async () => {
 				this.close();
-				const res = await Api.sendRequest("api/v1/tracks/remove-label/" + track.id + "/" + l.publicID, "POST");
+				const res = await Api.sendRequest("api/v1/tracks/remove-labels/" + track.id + "/" + l.publicID, "POST");
 				Popup.show(res.message || res.error, (res.code >= 200 && res.code <= 299 ? "success" : "error"), 5000);
 				await Api.get.tracks();
 				Api.show.tracks(Collection.filter());