+++ AAA_WEBAPP/assets/php/auth.php
@@ -1,3 +1,3 @@
 <?php
-header("Location: https://accounts.spotify.com/authorize?response_type=code&client_id=a209cbda1aaa4f408bd6ae2efc2264fb&redirect_uri=http%3A%2F%2Fspotify-labelling.21webb.nl%2Fassets%2Fphp%2Fget-access-code.php&scope=user-read-email%20user-read-private%20playlist-modify-public%20playlist-modify-private%20playlist-read-private%20playlist-read-collaborative%20user-library-read&show_dialog=false");
+header("Location: https://accounts.spotify.com/authorize?response_type=code&client_id=a209cbda1aaa4f408bd6ae2efc2264fb&redirect_uri=http%3A%2F%2Fspotify-labelling.21webb.nl%2Fassets%2Fphp%2Fget-access-code.php&scope=user-read-email%20playlist-modify-public%20playlist-modify-private%20playlist-read-private%20playlist-read-collaborative%20user-library-read&show_dialog=false");
 ?>